requirements.txt
nicegui>=1.4.0
requests

passlib[bcrypt]
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pytz>=2024.1
